Register Feature migration persistent task state named XContent (#84192)
This PR properly registers `NamedXContentRegistry` entries for `SystemIndexMigrationTaskParams` and `SystemIndexMigrationTaskState`. It also adds tests for the XContent de/serialization for those classes,
and fixes a bug revealed by these tests in `SystemIndexMigrationTaskState`'s parser.

Finally, it adds an integration test which simulates the conditions in which #84115 occurs: A node restart
while the migration is in progress. This ensures that we have fixed that particular bug.

(cherry picked from commit 4c3b7b1fb1cab30365df936a45d2a76142b4bf2d)
